requirements.txt
tweepy
requests
Pillow
python-dotenv